+++ core/src/com/unciv/logic/GameInfo.kt
@@ -10,7 +10,6 @@ import com.unciv.logic.BackwardCompatibility.ensureUnitIds
 import com.unciv.logic.BackwardCompatibility.guaranteeUnitPromotions
 import com.unciv.logic.BackwardCompatibility.migrateGreatGeneralPools
 import com.unciv.logic.BackwardCompatibility.migrateToTileHistory
-import com.unciv.logic.BackwardCompatibility.migrateUnitVariablesFromLegacyFields
 import com.unciv.logic.BackwardCompatibility.removeMissingModReferences
 import com.unciv.logic.GameInfoPreview.Companion.randomGameId
 import com.unciv.logic.automation.Timers.Companion.timeThis
@@ -911,7 +910,6 @@ class GameInfo : IsPartOfGameInfoSerialization, HasGameInfoSerializationVersion 
         guaranteeUnitPromotions()
         migrateToTileHistory()
         migrateGreatGeneralPools()
-        migrateUnitVariablesFromLegacyFields()
         ensureUnitIds()
     }