+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1261
-    const val appVersion = "4.21.10.6"
+    const val appCodeNumber = 1262
+    const val appVersion = "4.21.10.7"
     const val identifier = "com.unciv.app"
 }